fixed issues with enc dec
