+++ backend/Makefile
@@ -1,7 +1,5 @@
-setup: setup-nightly install-rocket
+setup: setup-nightly 
+	cargo run
 
 setup-nightly:
 	rustup override set nightly
-
-install-rocket:
-	cargo install rocket
