+++ backend/src/huffman/mod.rs
@@ -1,3 +1,4 @@
 pub mod coding;
 pub mod decoding;
 pub mod models;
+pub mod algorithms;
